% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_2d.R
\docType{package}
\name{PertGD-package}
\alias{PertGD}
\alias{PertGD-package}
\title{PertGD: Escaping the saddle points by perturbation}
\description{
The train_2d function is used to run the gradient-based
algorithms to find the stationary point in either first-order or second-order
sense.

Use the function \code{train_2d} to train your data.

See our \href{https://github.com/HongfanChen/PertGD}{Github home page}
or run ?train_2d for examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HongfanChen/PertGD}
  \item Report bugs at \url{https://github.com/HongfanChen/PertGD/issues}
}

}
\author{
\strong{Maintainer}: Hongfan Chen \email{chenhf@umich.edu}

Authors:
\itemize{
  \item Tunan Wang \email{wtunan@umich.edu}
}

}
