% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_2d.R
\name{train_2d}
\alias{train_2d}
\title{Performing faster perturbed gradient descent algorithm.}
\usage{
train_2d(trainer, params, gd, obj_f, epoch)
}
\arguments{
\item{trainer}{a character string specifying the algorithm performed.
Currently supported algorithms including: vanilla_gd, AGD, PAGD, FPGD, FPAGD.
See reference for detailed meaning.}

\item{params}{a list containing parameters required by the algorithm. See
examples for detail}

\item{gd}{a function specifying the explicit form of gradient function.}

\item{obj_f}{a function specifying the objective loss.}

\item{epoch}{integer. The number of iterations.}
}
\value{
A \strong{data.frame} consists of all the iterations.

The returned data.frame contains the following columns:
\itemize{
\item{Column 1: }{The first coordinate of iterate x}
\item{column 2: }{The second coordinate of iterate x}
\item{...}{...}
\item{column n:}{The n-th coordinate of iterate x}
}
}
\description{
Faster perturbed gradient descent algorithm designed for escaping the saddle
points more efficiently.
}
\examples{

params1 = list(x = c(0.6806,0), eta = 0.1, theta = 0.5, epsilon = 1e-2,
              radius = 1e-3, v = c(0,0), gamma = 1e-2, s = 0.1, count = 0,
              t = 10, zeta = 0, z = c(1,1), x_0 = c(1,1), iter = 0,
              t_sub = 5, eta_sub = 0.05)
train_2d(PAGD, params1, gd, obj_f, 20)

}
